Add the init_dependencies.main file to separate the imports and the main code for the dependencies
