+++ lib/init_dependencies.main.dart
@@ -1,0 +1,123 @@
+part of 'init_dependencies.dart';
+
+final serviceLocator = GetIt.instance;
+
+Future<void> initDependencies() async {
+  await _initSupabaseClient();
+  await _initHiveBox();
+  _initAuth();
+  _initBlogs();
+
+  serviceLocator.registerFactory(() => InternetConnection());
+
+  // Core
+  serviceLocator.registerLazySingleton(() => AppUserCubit());
+  serviceLocator.registerFactory<ConnectionChecker>(
+    () => ConnectionCheckerImpl(
+      serviceLocator(),
+    ),
+  );
+}
+
+Future<void> _initSupabaseClient() async {
+  await dotenv.load(fileName: ".env");
+  final supabase = await Supabase.initialize(
+    url: dotenv.env['SUPABASE_URL']!,
+    anonKey: dotenv.env['SUPABASE_ANON_KEY']!,
+  );
+  serviceLocator.registerLazySingleton<SupabaseClient>(() => supabase.client);
+}
+
+Future<void> _initHiveBox() async {
+  Hive.defaultDirectory = (await getApplicationDocumentsDirectory()).path;
+
+  serviceLocator.registerLazySingleton<Box>(
+    () => Hive.box(name: 'blogs'),
+  );
+}
+
+void _initAuth() {
+  // Repositories
+  serviceLocator.registerFactory<AuthRemoteDataSource>(
+    () => AuthRemoteDataSourceImpl(
+      serviceLocator(),
+    ),
+  );
+  serviceLocator.registerFactory<AuthRepository>(
+    () => AuthRepositoryImpl(
+      serviceLocator(),
+      serviceLocator(),
+    ),
+  );
+
+  // Use Cases
+  serviceLocator.registerFactory(
+    () => UserSignUp(
+      serviceLocator(),
+    ),
+  );
+  serviceLocator.registerFactory(
+    () => UserLogin(
+      serviceLocator(),
+    ),
+  );
+  serviceLocator.registerFactory(
+    () => UserSession(
+      serviceLocator(),
+    ),
+  );
+
+  // Blocs
+  serviceLocator.registerLazySingleton(
+    () => AuthBloc(
+      userSignUp: serviceLocator(),
+      userLogin: serviceLocator(),
+      userSession: serviceLocator(),
+      appUserCubit: serviceLocator(),
+    ),
+  );
+}
+
+void _initBlogs() {
+  // Data Source
+  serviceLocator.registerFactory<BlogRemoteDataSource>(
+    () => BlogRemoteDataSourceImpl(
+      serviceLocator(),
+    ),
+  );
+  serviceLocator.registerFactory<BlogLocalDataSource>(
+    () => BlogLocalDataSourceImpl(
+      serviceLocator(),
+    ),
+  );
+
+  // Repository
+  serviceLocator.registerFactory<BlogRepository>(
+    () => BlogRepositoryImpl(
+      serviceLocator(),
+      serviceLocator(),
+      serviceLocator(),
+    ),
+  );
+
+  // Use Case
+  serviceLocator.registerFactory(
+    () => UploadBlog(
+      serviceLocator(),
+    ),
+  );
+
+  serviceLocator.registerFactory(
+    () => FetchAllBlogs(
+      serviceLocator(),
+    ),
+  );
+
+  // Bloc
+  serviceLocator.registerLazySingleton(
+    () => BlogBloc(
+      uploadBlog: serviceLocator(),
+      fetchAllBlogs: serviceLocator(),
+    ),
+  );
+}
